freeze migrations into v26.6.2_0.ts, bump to 26.6.2:1
